Migrate repo from Poetry to uv tooling

+++ gaitmap/preprocessing/sensor_alignment/_gravity_alignment.py
@@ -82,7 +82,7 @@ def align_dataset_to_gravity(
     """
     dataset_type = is_sensor_data(dataset)
 
-    window_length = int(round(window_length_s * sampling_rate_hz))
+    window_length = round(window_length_s * sampling_rate_hz)
     acc_vector: Union[np.ndarray, dict[_Hashable, np.ndarray]]
     if dataset_type == "single":
         # get static acc vector
